+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Changed:
+* Magic Target Entities can now be turned invisible with potion effects.
+
 ### Fixed:
 * Optimized client performance by adjusting various things, like the trails rendering code (flowmotion / magnet blox)
 * MP Bar not rendering until you cast a spell first.